+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/ServerSocketChannelAcceptBlocked.kt
@@ -4,9 +4,8 @@ import org.pastalab.fray.core.concurrency.context.ServerSocketChannelContext
 import org.pastalab.fray.rmi.ResourceInfo
 import org.pastalab.fray.rmi.ResourceType
 
-class ServerSocketChannelAcceptBlocking(
-    val serverSocketChannelContext: ServerSocketChannelContext
-) : BlockedOperation(false, ResourceInfo(serverSocketChannelContext.port, ResourceType.SOCKET)) {
+class ServerSocketChannelAcceptBlocked(val serverSocketChannelContext: ServerSocketChannelContext) :
+    BlockedOperation(false, ResourceInfo(serverSocketChannelContext.port, ResourceType.NETWORK)) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     return null
   }