+++ instrumentation/base/src/main/kotlin/org/pastalab/fray/instrumentation/base/visitors/MethodExitVisitor.kt
@@ -6,6 +6,7 @@ import org.objectweb.asm.MethodVisitor
 import org.objectweb.asm.Opcodes
 import org.objectweb.asm.Type
 import org.objectweb.asm.commons.AdviceAdapter
+import org.pastalab.fray.runtime.Runtime
 
 class MethodExitVisitor(
     mv: MethodVisitor,
@@ -16,7 +17,9 @@ class MethodExitVisitor(
     val loadThis: Boolean,
     val loadArgs: Boolean,
     val addFinalBlock: Boolean,
-    val customizer: MethodExitVisitor.(v: MethodExitVisitor) -> Unit = {}
+    val customizer: MethodExitVisitor.(v: MethodExitVisitor, isFinalBlock: Boolean) -> Unit =
+        { v, isFinalBlock ->
+        }
 ) : AdviceAdapter(ASM9, mv, access, name, descriptor) {
   val methodEnterLabel = Label()
   val methodExitLabel = Label()
@@ -26,19 +29,23 @@ class MethodExitVisitor(
     super.visitCode()
   }
 
+  fun insertMethodExitBlock(isFinalBlock: Boolean) {
+    if (loadThis) {
+      loadThis()
+    }
+    if (loadArgs) {
+      loadArgs()
+    }
+    customizer(this, isFinalBlock)
+    invokeStatic(
+        Type.getObjectType(Runtime::class.java.name.replace(".", "/")),
+        Utils.kFunctionToASMMethod(method),
+    )
+  }
+
   override fun onMethodExit(opcode: Int) {
     if (opcode != ATHROW) {
-      if (loadThis) {
-        loadThis()
-      }
-      if (loadArgs) {
-        loadArgs()
-      }
-      customizer(this)
-      invokeStatic(
-          Type.getObjectType(org.pastalab.fray.runtime.Runtime::class.java.name.replace(".", "/")),
-          Utils.kFunctionToASMMethod(method),
-      )
+      insertMethodExitBlock(false)
     }
     super.onMethodExit(opcode)
   }
@@ -48,18 +55,8 @@ class MethodExitVisitor(
       visitLabel(methodExitLabel)
       catchException(methodEnterLabel, methodExitLabel, null)
       val locals = getLocals()
-      visitFrame(Opcodes.F_NEW, locals.size, getLocals(), 1, arrayOf("java/lang/Throwable"))
-      if (loadThis) {
-        loadThis()
-      }
-      if (loadArgs) {
-        loadArgs()
-      }
-      customizer(this)
-      invokeStatic(
-          Type.getObjectType(org.pastalab.fray.runtime.Runtime::class.java.name.replace(".", "/")),
-          Utils.kFunctionToASMMethod(method),
-      )
+      visitFrame(Opcodes.F_NEW, locals.size, locals, 1, arrayOf("java/lang/Throwable"))
+      insertMethodExitBlock(true)
       throwException()
     }
     super.visitMaxs(maxStack, maxLocals)