+++ integration-test/src/main/java/org/pastalab/fray/test/fail/network/SelectorWithNoRegistrationDeadlock.java
@@ -1,0 +1,15 @@
+package org.pastalab.fray.test.fail.network;
+
+import java.io.IOException;
+import java.nio.channels.Selector;
+
+public class SelectorWithNoRegistrationDeadlock {
+    public static void main(String[] args) throws IOException, InterruptedException {
+        Selector selector = Selector.open();
+        try {
+            selector.select();
+        } finally {
+            selector.close();
+        }
+    }
+}