+++ integration-test/src/test/java/org/pastalab/fray/test/FrayTestCase.java
@@ -13,10 +13,11 @@ import org.pastalab.fray.core.command.MethodExecutor;
 import org.pastalab.fray.core.randomness.ControlledRandom;
 import org.pastalab.fray.core.scheduler.PCTScheduler;
 import org.pastalab.fray.core.scheduler.POSScheduler;
-import org.pastalab.fray.test.fail.intstream.IntStream;
-import org.pastalab.fray.test.fail.thread.ThreadExitDeadlock;
-import org.pastalab.fray.test.success.network.SelectorTest;
-import org.pastalab.fray.test.success.thread.ThreadInterruptWithReentrantLockUnlock;
+import org.pastalab.fray.test.fail.network.AsyncClientSelectNoConnectDeadlock;
+import org.pastalab.fray.test.fail.network.SyncClientExceptionWithoutServer;
+import org.pastalab.fray.test.fail.network.SyncServerAcceptDeadlock;
+import org.pastalab.fray.test.success.network.AsyncServerAsyncClient;
+import org.pastalab.fray.test.success.network.SyncServerSyncClient;
 
 import java.util.*;
 
@@ -69,7 +70,7 @@ public class FrayTestCase {
                 new ExecutionInfo(
                         new LambdaExecutor(() -> {
                             try {
-                                SelectorTest.main(new String[]{});
+                                SyncClientExceptionWithoutServer.main(new String[]{});
                             } catch (Exception e) {
                                 throw new RuntimeException(e);
                             }