+++ runtime/src/main/java/org/pastalab/fray/runtime/Runtime.java
@@ -33,8 +33,7 @@ public class Runtime {
         DELEGATE.onThreadStartDone(t);
     }
 
-    // onThreadEnd and onThreadRun will only be called from JVM
-    // so no recursion check is necessary.
+    // onThreadEnd and onThreadRun will only be called from JVM.
     public static void onThreadEnd() {
         DELEGATE.onThreadEnd();
     }
@@ -487,36 +486,44 @@ public class Runtime {
         DELEGATE.onRangerCondition(condition);
     }
 
+
     public static void onSelectorSelect(Selector selector) {
-    }
-
-    public static void onSelectorSelectDone(Selector selector) {
-    }
-
-    public static void onSelectorRegister(Selector selector, SelectionKey key) {
-    }
-
-    public static void onSelectorCancel(Selector selector, SelectionKey key) {
-    }
-
-    public static void onSelectorSetEventOps(Selector selector, SelectionKey key) {
+        DELEGATE.onSelectorSelect(selector);
     }
 
     public static void onServerSocketChannelBindDone(ServerSocketChannel channel) {
+        DELEGATE.onServerSocketChannelBindDone(channel);
     }
 
     public static void onServerSocketChannelAccept(ServerSocketChannel channel) {
+        DELEGATE.onServerSocketChannelAccept(channel);
     }
 
-    public static void onServerSocketChannelAcceptDone(ServerSocketChannel channel) {
+    public static void onSocketChannelConnect(SocketChannel channel, SocketAddress remoteAddress) {
+        DELEGATE.onSocketChannelConnect(channel, remoteAddress);
     }
 
-    public static void onSocketChannelWrite(SocketChannel channel) {
+    public static void onSelectorSetEventOpsDone(Selector selector, SelectionKey key) {
+        DELEGATE.onSelectorSetEventOpsDone(selector, key);
     }
 
-    public static void onSocketChannelRead(SocketChannel channel) {
+    public static void onSelectorCancelKeyDone(Selector selector, SelectionKey key) {
+        DELEGATE.onSelectorCancelKeyDone(selector, key);
     }
 
-    public static void onSocketChannelConnect(SocketChannel channel, SocketAddress address) {
+    public static void onSelectorSelectDone(Selector selector) {
+        DELEGATE.onSelectorSelectDone(selector);
+    }
+
+    public static void onServerSocketChannelAcceptDone(ServerSocketChannel channel, SocketChannel client) {
+        DELEGATE.onServerSocketChannelAcceptDone(channel, client);
+    }
+
+    public static void onSocketChannelCloseDone(SocketChannel channel) {
+        DELEGATE.onSocketChannelCloseDone(channel);
+    }
+
+    public static void onSocketChannelConnectDone(SocketChannel channel, boolean success) {
+        DELEGATE.onSocketChannelConnectDone(channel, success);
     }
 }